LayoutTests/plugins/npruntime/object-from-destroyed-plugin-in-subframe-expected.txt

Test various operation on an NPObject whose plug-in (that lives in a subframe) has been destroyed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


 PASS testObject.gettingProperty threw exception ReferenceError: Trying to access object from destroyed plug-in..
PASS testObject.settingProperty = 10 threw exception ReferenceError: Trying to access object from destroyed plug-in..
PASS testObject() threw exception TypeError: '' is not a function (evaluating 'testObject()').
PASS new testObject(); threw exception TypeError: '' is not a constructor (evaluating 'new testObject()').
